Move processing to seperate package
