+++ src/ner_telhub/model/data_models.py
@@ -9,8 +9,8 @@ from PyQt6.QtCore import (
     pyqtSignal, pyqtBoundSignal
 )
 
-from ner_telhub.model.processing.master_mapping import MESSAGE_IDS, DECODE_DATA, DATA_IDS
-from ner_telhub.model.processing.data import Data
+from ner_processing.master_mapping import MESSAGE_IDS, DECODE_DATA, DATA_IDS
+from ner_processing.data import Data
 from ner_telhub.utils.threads import Worker
